fixed 'thread not exist' bug in change_priority function

+++ src/thread/scheduler/priority.rs
@@ -43,19 +43,20 @@ impl Schedule for Priority {
 
     fn change_priority(&mut self, thread: Arc<Thread>, priority: u32) {
         let tid = thread.id();
-        let old_priority = *self.thread_to_priority.get(&tid).unwrap();
-        self.thread_to_priority
-            .entry(tid)
-            .and_modify(|v| *v = priority);
+        if let Some(old_priority) = self.thread_to_priority.get(&tid).copied() {
+            self.thread_to_priority
+                .entry(tid)
+                .and_modify(|v| *v = priority);
 
-        let old_bucket = self.priority_to_thread.get_mut(&old_priority).unwrap();
-        if let Some(index) = old_bucket.iter().position(|x| Arc::ptr_eq(x, &thread)) {
-            old_bucket.remove(index);
+            let old_bucket = self.priority_to_thread.get_mut(&old_priority).unwrap();
+            if let Some(index) = old_bucket.iter().position(|x| Arc::ptr_eq(x, &thread)) {
+                old_bucket.remove(index);
+            }
+
+            self.priority_to_thread
+                .entry(priority)
+                .or_default()
+                .push_back(thread);
         }
-
-        self.priority_to_thread
-            .entry(priority)
-            .or_default()
-            .push_back(thread);
     }
 }
